Fix imports after move of live_os payload
