+++ pyanaconda/modules/payload/payloads/live_os/live_os.py
@@ -30,7 +30,7 @@ from pyanaconda.modules.payload.base.handler_base import PayloadHandlerBase
 from pyanaconda.modules.payload.base.initialization import PrepareSystemForInstallationTask, \
     CopyDriverDisksFilesTask, SetUpSourcesTask, TearDownSourcesTask
 from pyanaconda.modules.payload.base.utils import get_dir_size
-from pyanaconda.modules.payload.live.live_os_interface import LiveOSHandlerInterface
+from pyanaconda.modules.payload.payloads.live_os.live_os_interface import LiveOSHandlerInterface
 from pyanaconda.modules.payload.live.initialization import UpdateBLSConfigurationTask
 from pyanaconda.modules.payload.live.installation import InstallFromImageTask
 from pyanaconda.modules.payload.live.utils import get_kernel_version_list
